I've made some progress on adding the capability to grab one dark buffer than ask the user to select a threshold from our pulse histogram.
I'm running into a problem with connecting the ROOT Canvas/TPad/etc internal ProcessEvents to an external function. Right now it looks like I'll have to write a simple
ROOT class and generate a dictionary for it.

Ideally I'd like to avoid that entirely, or at least generate a simple dictionary connection for the function I'd like to use (this approach doesn't appear to work).